refactor: decouple default service name from the versioned schematize_service_name function (#19030)
This change eliminates 2500 import cycles by decoupling the default
service named relied on by `ddtrace.settings` from the versioned
`schematize_service_name` function definition relied on by most
contribs.

(cherry picked from commit 727771442208e04aec8293befd6826a3dd7d4e45)
