+++ tests/contrib/aiohttp/test_middleware.py
@@ -72,7 +72,7 @@ import pytest
 import asyncio
 from tests.conftest import *
 from tests.contrib.aiohttp.conftest import *
-from ddtrace.internal.schema import DEFAULT_SPAN_SERVICE_NAME
+from ddtrace.internal.schema.default import DEFAULT_SPAN_SERVICE_NAME
 
 def test(app, loop, aiohttp_client, test_spans):
     async def async_test(app, aiohttp_client, test_spans):
