+++ tests/contrib/pymysql/test_pymysql.py
@@ -3,7 +3,7 @@ import pymysql
 
 from ddtrace.contrib.internal.pymysql.patch import patch
 from ddtrace.contrib.internal.pymysql.patch import unpatch
-from ddtrace.internal.schema import DEFAULT_SPAN_SERVICE_NAME
+from ddtrace.internal.schema.default import DEFAULT_SPAN_SERVICE_NAME
 from tests.contrib import shared_tests
 from tests.utils import TracerTestCase
 from tests.utils import assert_dict_issuperset
