+++ tests/contrib/redis/test_redis_cluster_asyncio.py
@@ -246,7 +246,7 @@ def test_default_service_name_v1():
     import redis
 
     from ddtrace.contrib.internal.redis.patch import patch
-    from ddtrace.internal.schema import DEFAULT_SPAN_SERVICE_NAME
+    from ddtrace.internal.schema.default import DEFAULT_SPAN_SERVICE_NAME
     from tests.contrib.config import REDISCLUSTER_CONFIG
     from tests.utils import TracerSpanContainer
     from tests.utils import scoped_tracer
