+++ tests/contrib/urllib3/test_urllib3.py
@@ -10,7 +10,7 @@ from ddtrace.constants import ERROR_TYPE
 from ddtrace.contrib.internal.urllib3.patch import patch
 from ddtrace.contrib.internal.urllib3.patch import unpatch
 from ddtrace.ext import http
-from ddtrace.internal.schema import DEFAULT_SPAN_SERVICE_NAME
+from ddtrace.internal.schema.default import DEFAULT_SPAN_SERVICE_NAME
 from tests.contrib.config import HTTPBIN_CONFIG
 from tests.utils import TracerTestCase
 from tests.utils import snapshot
