+++ tests/contrib/valkey/test_valkey_cluster_asyncio.py
@@ -175,7 +175,7 @@ def test_default_service_name_v1():
     import valkey
 
     from ddtrace.contrib.internal.valkey.patch import patch
-    from ddtrace.internal.schema import DEFAULT_SPAN_SERVICE_NAME
+    from ddtrace.internal.schema.default import DEFAULT_SPAN_SERVICE_NAME
     from tests.contrib.config import VALKEY_CLUSTER_CONFIG
     from tests.utils import TracerSpanContainer
     from tests.utils import scoped_tracer
